requirements.txt
pyquery==1.2.9
requests==2.7.0
hurry.filesize==0.9